refactor: use fetchFromGitHub as the GitHub fetcher (#30)
* refactor: use fetchFromGitHub as the GitHub fetcher

This change aims to save bandwidth when fetching large repositories from
GitHub. It does this by using fetchFromGitHub instead of fetchgit.
fetchFromGitHub, wherever possible, relies on fetchzip to download the
archive of the desired revision instead of downloading the entire git
history.

As part of this attempt, the git prefetcher is replaced from
nix-prefetch-git to nix-prefetch[1]. The primary purpose for this change
is wider and improved support for Nix prefetchers. nix-prefetch allows nvfetcher
to support a wider range of Nix fetcher functions, including fetchFromGitHub.
It can also work with any argument that those Nix fetchers take as input.

[1]: https://github.com/msteen/nix-prefetch/

* fixup! refactor: use fetchFromGitHub as the GitHub fetcher

* Update flake.lock

* fixup! refactor: use fetchFromGitHub as the GitHub fetcher

* Disable aarch64-darwin build temporarily to deal with GHC breakage

* fixup! refactor: use fetchFromGitHub as the GitHub fetcher

Co-authored-by: Potato Hatsue <berberman@yandex.com>

* fixup! refactor: use fetchFromGitHub as the GitHub fetcher

Co-authored-by: Potato Hatsue <berberman@yandex.com>

* Update source location for QLivePlayer in examples

Co-authored-by: Potato Hatsue <berberman@yandex.com>